refactor : 역할생성 및 책임 이동
